Generalize ok/fail logic and implement config file
Instead of hardcoding the meaning of output on stdout, stderr and the
return code, match output on these three streams (the return code is
treated as another stream consisting of a single line) against regexps
and implement a simple 4 level ok/fail+sticky schema.

If the argument is a diretory, it is supposed to contain a config.json
file and any output is also written to that directory.