Add config option min_free_pct (default 1)
Cronwrapper is often used to wrap jobs which can potentially create lots
of output and can fill up a disk faster than the failure can be
detected (or at least fixed). Therefore it is useful to have an
emergency brake which refuses to run the job if the disk is almost full.

Cronwrapper will still write a log file if that happens but it will
contain only a single line. Assuming a 1 GB filesystem with a 4kB block
size, the default of 1 % means that anouther 2560 logs can be written
before the file system is really full. At one invocation per minute that
would be 42+ hours.